DEVAS-BM-helper: add sharePostButton story
