+++ assembl/static2/.storybook/config.js
@@ -35,6 +35,7 @@ function loadStories() {
 
   require('../js/app/stories/components/debate/common/toggleCommentButton.stories.jsx');
   require('../js/app/stories/components/debate/common/replyToCommentButton.stories.jsx');
+  require('../js/app/stories/components/debate/common/sharePostButton.stories.jsx');
 
   require('../js/app/stories/components/debate/brightMirror/fictionPreview.stories.jsx');
   require('../js/app/stories/components/debate/brightMirror/fictionsList.stories.jsx');
